requirements.txt
django==4.2.5
django-otp==1.2.3
cryptography==41.0.4
pyotp==2.9.0
qrcode==7.4.2 
pillow==10.0.1
vonage==3.10.0
python-dotenv==1.0.0

